requirements.txt
asgiref==3.2.10
Django==2.2.1
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
